[FEAT] 마이핏 페이지 UI 구현 및 API 연동 (#12)

+++ app/src/main/java/com/refit/app/data/myfit/model/UpdateMemberProductRequest.kt
@@ -4,7 +4,7 @@ data class UpdateMemberProductRequest(
     val productName: String? = null,
     val brandName: String? = null,
     val categoryId: Long? = null,
-    val effectIds: List<Int>? = null,
+    val effectIds: List<Long>? = null,
     val recommendedPeriod: Int? = null,
     val startDate: String? = null // YYYY-MM-DD
 )
